requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.2
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
text-unidecode==1.3
twilio==9.6.5
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
